Rename rule to no-primitive-constructors
