+++ .eslintrc.js
@@ -55,7 +55,7 @@ module.exports = {
     // CUSTOM RULES
     // the second argument of warning/invariant should be a literal string
     'react-internal/warning-and-invariant-args': ERROR,
-    'react-internal/boolean-and-string-constructors': ERROR,
+    'react-internal/no-primitive-constructors': ERROR,
   },
 
   globals: {
