+++ eslint-rules/index.js
@@ -3,6 +3,6 @@
 module.exports = {
   rules: {
     'warning-and-invariant-args': require('./warning-and-invariant-args'),
-    'boolean-and-string-constructors': require('./boolean-and-string-constructors'),
+    'no-primitive-constructors': require('./no-primitive-constructors'),
   },
 };
