+++ eslint-rules/no-primitive-constructors.js
